DIP-bot/requirements.txt
python-binance
telegram_send
schedule
gspread
requests
json